+++ public/src/app/basket/basket.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { OmniService } from '.././omni.service';
 
 @Component({
